Refactor: move SubDiv class to models
