+++ repondeur/zam_repondeur/models/division.py
@@ -1,0 +1,8 @@
+from typing import NamedTuple
+
+
+class SubDiv(NamedTuple):
+    type_: str
+    num: str
+    mult: str
+    pos: str
